editregions/contrib/textfiles/requirements.txt
django-template-finder>=0.5
CommonMark>=0.5.4
